+++ src/bot/templates/src/app/scripts/{botName}/{staticTabClassName}Tab.tsx
@@ -1,13 +1,5 @@
 import * as React from "react";
-import {
-    Panel,
-    PanelBody,
-    PanelHeader,
-    PanelFooter,
-    Surface,
-    TeamsThemeContext,
-    getContext
-} from "msteams-ui-components-react";
+import { Provider, Flex, Text, Button, Header } from "@stardust-ui/react";
 import TeamsBaseComponent, { ITeamsBaseComponentProps, ITeamsBaseComponentState } from "msteams-react-base-component";
 import * as microsoftTeams from "@microsoft/teams-js";
 
@@ -46,37 +38,26 @@ export class <%=staticTabClassName%>Tab extends TeamsBaseComponent<I<%=staticTab
      * The render() method to create the UI of the tab
      */
     public render() {
-        const context = getContext({
-            baseFontSize: this.state.fontSize,
-            style: this.state.theme
-        });
-        const { rem, font } = context;
-        const { sizes, weights } = font;
-        const styles = {
-            header: { ...sizes.title, ...weights.semibold },
-            section: { ...sizes.base, marginTop: rem(1.4), marginBottom: rem(1.4) },
-            footer: { ...sizes.xsmall }
-        };
         return (
-            <TeamsThemeContext.Provider value={context}>
-                <Surface>
-                    <Panel>
-                        <PanelHeader>
-                            <div style={styles.header}>Welcome to the <%= botTitle%> bot page</div>
-                        </PanelHeader>
-                        <PanelBody>
-                            <div style={styles.section}>
-                                TODO: Add your content here
-                            </div>
-                        </PanelBody>
-                        <PanelFooter>
-                            <div style={styles.footer}>
-                                (C) Copyright <%=developer%>
-                            </div>
-                        </PanelFooter>
-                    </Panel>
-                </Surface>
-            </TeamsThemeContext.Provider>
+            <Provider theme={this.state.theme}>
+                <Flex fill={true} column styles={{
+                    padding: ".8rem 0 .8rem .5rem"
+                }}>
+                    <Flex.Item>
+                        <Header content="Welcome to the <%= botTitle%> bot page" />
+                    </Flex.Item>
+                    <Flex.Item>
+                        <div>
+                            <Text content="TODO: Add you content here" />
+                        </div>
+                    </Flex.Item>
+                    <Flex.Item styles={{
+                        padding: ".8rem 0 .8rem .5rem"
+                    }}>
+                        <Text size="smaller" content="(C) Copyright <%=developer%>" />
+                    </Flex.Item>
+                </Flex>
+            </Provider>
         );
     }
 }