+++ src/connector/ConnectorGenerator.ts
@@ -123,7 +123,7 @@ export class ConnectorGenerator extends Generator {
             Yotilities.addAdditionalDeps([
                 ['@types/node-json-db', '0.0.1'],
                 ['node-json-db', '0.7.5'],
-                ["msteams-ui-components-react", "^0.8.1"],
+                ["@stardust-ui/react", "^0.38.1"],
                 ["react", "^16.8.4"],
                 ["@types/react", "16.8.8"],
                 ["react-dom", "^16.8.4"],