+++ src/messageExtension/templates/src/app/scripts/{messageExtensionName}/{messageExtensionClassName}Action.tsx
@@ -1,15 +1,5 @@
 import * as React from "react";
-import {
-    PrimaryButton,
-    Panel,
-    PanelBody,
-    PanelHeader,
-    PanelFooter,
-    Surface,
-    Input,
-    TeamsThemeContext,
-    getContext
-} from "msteams-ui-components-react";
+import { Provider, Flex, Header, Input, Button, Text } from "@stardust-ui/react";
 import TeamsBaseComponent, { ITeamsBaseComponentProps, ITeamsBaseComponentState } from "msteams-react-base-component";
 import * as microsoftTeams from "@microsoft/teams-js";
 
@@ -46,55 +36,35 @@ export class <%=messageExtensionClassName%>Action extends TeamsBaseComponent<I<%
      * The render() method to create the UI of the tab
      */
     public render() {
-        const context = getContext({
-            baseFontSize: this.state.fontSize,
-            style: this.state.theme
-        });
-        const { rem, font } = context;
-        const { sizes, weights } = font;
-        const styles = {
-            header: { ...sizes.title, ...weights.semibold },
-            section: { ...sizes.base, marginTop: rem(1.4), marginBottom: rem(1.4) },
-            footer: { ...sizes.xsmall }
-        };
         return (
-            <TeamsThemeContext.Provider value={context}>
-                <Surface>
-                    <Panel>
-                        <PanelHeader>
-                            <div style={styles.header}><%= messageExtensionTitle%> configuration</div>
-                        </PanelHeader>
-                        <PanelBody>
-                            <div style={styles.section}>
-                                 <Input
-                                    autoFocus
-                                    placeholder="email@email.com"
-                                    label="Enter an e-mail address"
-                                    errorLabel={!this.state.email ? "This value is required" : undefined}
-                                    value={this.state.email}
-                                    onChange={(e) => {
+            <Provider theme={this.state.theme}>
+                <Flex fill={true} column styles={{
+                    padding: ".8rem 0 .8rem .5rem"
+                }}>
+                    <Flex.Item>
+                        <div>
+                            <Header content="<%= messageExtensionTitle%> configuration" />
+                            <Text content="Enter an e-mail address" />
+                            <Input
+                                placeholder="email@email.com"
+                                fluid
+                                clearable
+                                value={this.state.email}
+                                onChange={(e, data) => {
+                                    if (data) {
                                         this.setState({
-                                            email: e.target.value
+                                            email: data.value
                                         });
-                                    }}
-                                    required />
-                            </div>
-                            <div style={styles.section}>
-                                <PrimaryButton onClick={() => {
-                                    microsoftTeams.tasks.submitTask({
-                                        email: this.state.email
-                                    });
-                                }}>OK</PrimaryButton>
-                            </div>
-                        </PanelBody>
-                        <PanelFooter>
-                            <div style={styles.footer}>
-                                (C) Copyright <%=developer%>
-                            </div>
-                        </PanelFooter>
-                    </Panel>
-                </Surface>
-             </TeamsThemeContext.Provider>
+                                    }
+                                }}
+                                required />
+                            <Button onClick={() => microsoftTeams.tasks.submitTask({
+                                    email: this.state.email
+                                })} primary>OK</Button>
+                        </div>
+                    </Flex.Item>
+                </Flex>
+            </Provider>
         );
     }
 }