+++ src/tab/TabGenerator.ts
@@ -190,7 +190,7 @@ export class TabGenerator extends Generator {
             this.fs.writeJSON(manifestPath, manifest);
 
             Yotilities.addAdditionalDeps([
-                ["msteams-ui-components-react", "^0.8.1"],
+                ["@stardust-ui/react", "^0.38.1"],
                 ["react", "^16.8.4"],
                 ["@types/react", "16.8.8"],
                 ["react-dom", "^16.8.4"],